+++ fabric/src/main/java/smartin/miapi/fabric/MiapiClientFabric.java
@@ -5,6 +5,7 @@ import net.fabricmc.api.Environment;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import net.fabricmc.fabric.api.client.rendering.v1.ArmorRenderer;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.UseAnim;
 import smartin.miapi.client.model.item.ItemBakedModelReplacement;
 import smartin.miapi.registries.RegistryInventory;
 
@@ -14,6 +15,8 @@ import java.util.List;
 public class MiapiClientFabric {
 
     public static void setupClient() {
+        UseAnim anim;
+
         //smartin.miapi.client.MiapiClient.KEY_BINDINGS.addCallback((KeyBindingHelper::registerKeyBinding));
         RegistryInventory.MODULAR_ITEMS.addCallback((item) -> ArmorRenderer.register(new ModularArmorRenderer(), item));
         ModelLoadingPlugin.register(pluginContext -> {